Splitting up code & renaming to fit conventions
